+++ docs/changelog.rst
@@ -11,6 +11,8 @@ Features
   all listed in :mod:`pycaputo.fode.gallery`.
 * Add some neuron models: FitzHugh-Rinzel. This is on top of the existing
   Integrate-and-Fire models from :mod:`pycaputo.integrate_fire`.
+* Made :class:`~pycaputo.fode.caputo.Trapezoidal` a bit faster by avoiding
+  multiple ``einsum``.
 
 Breaking Changes
 ^^^^^^^^^^^^^^^^